Integ: trilead-ssh2 build-217-jenkins-11 (#22250)

+++ nucleus/cluster/admin/src/main/java/com/sun/enterprise/v3/admin/cluster/SecureAdminBootstrapHelper.java
@@ -313,8 +313,8 @@ public abstract class SecureAdminBootstrapHelper {
          * @param milliseconds normal Java time (in milliseconds)
          * @return
          */
-        Integer secondsSince_01_Jan_1970(final long milliseconds) {
-            return Integer.valueOf((int) (milliseconds) / 1000);
+        Long secondsSince_01_Jan_1970(final long milliseconds) {
+            return milliseconds / 1000;
         }
     }
